+++ src/metaheuristics/LocalSearch.java
@@ -46,7 +46,7 @@ public class LocalSearch extends Solver{
 	 * @see metaheuristics.Solver#solve(util.Timer)
 	 */
 	public void solve(Timer timer) {
-		int cmax = this.getSolution().getScore();
+		double cmax = this.getSolution().getScore();
 		// la recherche dans le voisinage est faite de sorte à ne pas dégrader
 		// la solution lorsqu'une cuvette est atteinte, donc la boucle suivante est correcte.
 		do{
